Fix power LED aliasing
The power LED GPIO is put in the power GPIO array and initialised
separately from the led0,1,2 initialisation. However they alias the
same device. This causes led_init() to clobber the power-on state
of the power LED which makes the Nucleo look like it powered on
without the host (green LED) powered on.

Instead of having this aliasing, require the board overlay define
a status led and power and reset gpios (which may be LEDs). Then
these are used independently.

The led_toggle shell command is removed since it's outlived its
usefulness for testing.