jtag: Make kconfig gracefully handle lack of jtag
When the device tree nodes are missing the build fails due to asserts in
the jtag code.

Make the JTAG option depend on the presence of a jtag node. This could
be improved when the jtag code gets it's own compatible, and we could
search for that (or a label), which would be less fragile.