Add RTC support
This adds initial RTC support. The Nucleo STM32 has an integrated RTC,
which is used to set system time at boot.

So long as the board is powered, the RTC keeps time across resets. P550
should work too, and it has a button battery.

Currently there is no way to set the time. The builtin 'date set' command
sets the system time but not the RTC, so disable that for now.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>