Fix and re-enable persistent storage on the Nucleo
This fixes the partition scheme so that the program image is not
overwritten by application storage, which seems to get it working
well enough to enable.

This likely still requires a Zephyr patch to avoid XIP while writing
flash to be robust, but it seems to work without it. The Zephyr patch
is in progress.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>