Add authentication for console websocket
Add authentication for the console websocket. This can't
use the same Authorization: header that redfish uses
because the WebSocket connection API can not send headers
with the upgrade request. A special Sec-WebSocket-Protocol
header can be used in a pinch.

This doesn't use base64 encoding because that's useless
and also this ws header has some validation which disallows
various special characters like : and =. When HTTPS is used,
authentication should be secure.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>