Reduce config to under 64 bytes
The littlefs filesystem requires a sector for filesystem structures,
one for file metadata, and one for file data. The P550 flash
arrangement can only provide 2 sectors for application storage.

Littlefs can store data inline in the file/inode metadata if it
is <= 64 bytes, and does not require the 3rd sector. Squash the
config down to under 64 bytes and bump the version. This should
keep persistent storage for P550 limping along for a while.